Break up tags code from category interface (#28439)
